requirement.txt
crewai
tavily-python
openai
langchain
pydantic
python-dotenv
